+++ saleor/core/telemetry/trace.py
@@ -1,27 +1,29 @@
 import logging
-from abc import ABC, abstractmethod
 from collections.abc import Iterator, Sequence
-from contextlib import AbstractContextManager, contextmanager
-from typing import TypedDict
+from contextlib import contextmanager
 
-from opentelemetry.trace import INVALID_SPAN, Link, Span, SpanKind
+from opentelemetry.trace import (
+    INVALID_SPAN,
+    Link,
+    Span,
+    SpanKind,
+    get_current_span,
+    get_tracer,
+)
 from opentelemetry.util.types import Attributes
 
-from .utils import enrich_with_global_attributes
+from ... import __version__ as saleor_version
+from .utils import CORE_SCOPE, SERVICE_SCOPE, enrich_with_global_attributes
 
 logger = logging.getLogger(__name__)
 
 
-class SpanConfig(TypedDict):
-    kind: SpanKind
-    attributes: Attributes
-    links: Sequence[Link] | None
-    start_time: int | None
-    record_exception: bool
-    set_status_on_exception: bool
+class Tracer:
+    def __init__(self):
+        self._core_tracer = get_tracer(CORE_SCOPE, saleor_version)
+        self._service_tracer = get_tracer(SERVICE_SCOPE, saleor_version)
 
-
-class Tracer(ABC):
+    @contextmanager
     def start_as_current_span(
         self,
         name: str,
@@ -34,25 +36,20 @@ class Tracer(ABC):
         record_exception: bool = True,
         set_status_on_exception: bool = True,
         end_on_exit: bool = True,
-    ) -> AbstractContextManager[Span]:
+    ) -> Iterator[Span]:
         attributes = enrich_with_global_attributes(attributes)
-        span_config = SpanConfig(
+        tracer = self._service_tracer if service_scope else self._core_tracer
+        with tracer.start_as_current_span(
+            name,
             kind=kind,
             attributes=attributes,
             links=links,
             start_time=start_time,
             record_exception=record_exception,
             set_status_on_exception=set_status_on_exception,
-        )
-        return self._start_as_current_span(
-            name, service_scope, span_config, end_on_exit
-        )
-
-    @abstractmethod
-    def _start_as_current_span(
-        self, name: str, service_scope: bool, span_config: SpanConfig, end_on_exit: bool
-    ) -> AbstractContextManager[Span]:
-        pass
+            end_on_exit=end_on_exit,
+        ) as span:
+            yield span
 
     def start_span(
         self,
@@ -67,7 +64,9 @@ class Tracer(ABC):
         set_status_on_exception: bool = True,
     ) -> Span:
         attributes = enrich_with_global_attributes(attributes)
-        span_config = SpanConfig(
+        tracer = self._service_tracer if service_scope else self._core_tracer
+        return tracer.start_span(
+            name,
             kind=kind,
             attributes=attributes,
             links=links,
@@ -75,17 +74,9 @@ class Tracer(ABC):
             record_exception=record_exception,
             set_status_on_exception=set_status_on_exception,
         )
-        return self._start_span(name, service_scope, span_config)
 
-    @abstractmethod
-    def _start_span(
-        self, name: str, service_scope: bool, span_config: SpanConfig
-    ) -> Span:
-        pass
-
-    @abstractmethod
     def get_current_span(self) -> Span:
-        pass
+        return get_current_span()
 
 
 class TracerProxy(Tracer):
@@ -98,17 +89,59 @@ class TracerProxy(Tracer):
         self._tracer = tracer_cls()
 
     @contextmanager
-    def _start_as_current_span(self, *args, **kwargs) -> Iterator[Span]:
+    def start_as_current_span(
+        self,
+        name: str,
+        *,
+        service_scope: bool = False,
+        kind: SpanKind = SpanKind.INTERNAL,
+        attributes: Attributes = None,
+        links: Sequence[Link] | None = None,
+        start_time: int | None = None,
+        record_exception: bool = True,
+        set_status_on_exception: bool = True,
+        end_on_exit: bool = True,
+    ) -> Iterator[Span]:
         if self._tracer is None:
             yield INVALID_SPAN
         else:
-            with self._tracer._start_as_current_span(*args, **kwargs) as span:
+            with self._tracer.start_as_current_span(
+                name,
+                service_scope=service_scope,
+                kind=kind,
+                attributes=attributes,
+                links=links,
+                start_time=start_time,
+                record_exception=record_exception,
+                set_status_on_exception=set_status_on_exception,
+                end_on_exit=end_on_exit,
+            ) as span:
                 yield span
 
-    def _start_span(self, *args, **kwargs) -> Span:
+    def start_span(
+        self,
+        name: str,
+        *,
+        service_scope: bool = False,
+        kind: SpanKind = SpanKind.INTERNAL,
+        attributes: Attributes = None,
+        links: Sequence[Link] | None = None,
+        start_time: int | None = None,
+        record_exception: bool = True,
+        set_status_on_exception: bool = True,
+    ) -> Span:
         if self._tracer is None:
             return INVALID_SPAN
-        return self._tracer._start_span(*args, **kwargs)
+        return self._tracer.start_span(
+            name,
+            service_scope=service_scope,
+            kind=kind,
+            attributes=attributes,
+            links=links,
+            start_time=start_time,
+            record_exception=record_exception,
+            set_status_on_exception=set_status_on_exception,
+        )
 
     def get_current_span(self) -> Span:
         if self._tracer is None:
