+++ saleor/settings.py
@@ -968,5 +968,5 @@ TRANSACTION_ITEMS_LIMIT = 100
 # memcached to avoid leaking key values.
 warnings.filterwarnings("ignore", category=CacheKeyWarning)
 
-TELEMETRY_TRACER_CLASS = "saleor.core.telemetry.otel.OpenTelemetryTracer"
-TELEMETRY_METER_CLASS = "saleor.core.telemetry.otel.OpenTelemetryMeter"
+TELEMETRY_TRACER_CLASS = "saleor.core.telemetry.trace.Tracer"
+TELEMETRY_METER_CLASS = "saleor.core.telemetry.metric.Meter"
